fix 2

+++ crates/graphdb-core/src/core/types/index.rs
@@ -217,7 +217,7 @@ impl IndexField {
 /// Indexed Configuration Structures
 #[derive(Debug, Clone)]
 pub struct IndexConfig {
-    pub id: i32,
+    pub id: u64,
     pub name: String,
     pub space_id: u64,
     pub schema_name: String,
@@ -231,7 +231,7 @@ pub struct IndexConfig {
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct Index {
-    pub id: i32,
+    pub id: u64,
     pub name: String,
     pub space_id: u64,
     pub schema_name: String,
